:recycle: 캣관련 컴포넌트 스타일 분리 및 구조 변경
